model fix
